Move some drilled props to a context
- Move inputAddress, tokenMapping, openSeaProxyAddress, settings and
  setInputAddress, setSettings to useAppContext hook
- Move around code for higher levels of the Dashboard component
- Move around some other code
