+++ components/Dashboard/Allowance.tsx
@@ -1,4 +1,5 @@
 import AllowanceControls from 'components/common/AllowanceControls';
+import { useAppContext } from 'lib/hooks/useAppContext';
 import { useEthereum } from 'lib/hooks/useEthereum';
 import { useRevoke } from 'lib/hooks/useRevoke';
 import { ITokenAllowance, TokenData } from 'lib/interfaces';
@@ -9,15 +10,14 @@ import { ClipLoader } from 'react-spinners';
 import AllowanceDisplay from './AllowanceDisplay';
 
 interface Props {
-  inputAddress: string;
   token: TokenData;
   allowance: ITokenAllowance;
   onRevoke: (allowance: ITokenAllowance) => void;
-  openSeaProxyAddress?: string;
 }
 
-function Allowance({ allowance, inputAddress, token, onRevoke, openSeaProxyAddress }: Props) {
+function Allowance({ allowance, token, onRevoke }: Props) {
   const { selectedChainId } = useEthereum();
+  const { openSeaProxyAddress } = useAppContext();
   const { revoke, update, updatedAmount } = useRevoke(token, allowance, onRevoke);
   const { result: spenderName, loading } = useAsync(
     () => addressToAppName(allowance.spender, selectedChainId, openSeaProxyAddress),
@@ -35,12 +35,7 @@ function Allowance({ allowance, inputAddress, token, onRevoke, openSeaProxyAddre
   return (
     <Form inline className="Allowance">
       <AllowanceDisplay token={token} allowance={allowance} updatedAmount={updatedAmount} spenderName={spenderName} />
-      <AllowanceControls
-        revoke={revoke}
-        update={update}
-        inputAddress={inputAddress}
-        id={`${token.contract.address}-${allowance.spender}`}
-      />
+      <AllowanceControls revoke={revoke} update={update} id={`${token.contract.address}-${allowance.spender}`} />
     </Form>
   );
 }
