+++ components/Dashboard/Dashboard.tsx
@@ -1,35 +1,16 @@
-import { EthereumProvider } from 'lib/hooks/useEthereum';
-import { ToastContainer } from 'react-toastify';
 import DashboardBody from './DashboardBody';
+import ConnectSection from './header/ConnectSection';
 import DashboardHeader from './header/DashboardHeader';
-
-const SafeHydrate = ({ children }) => {
-  return <div suppressHydrationWarning>{typeof window === 'undefined' ? null : children}</div>;
-};
 
 function Dashboard() {
   return (
-    <SafeHydrate>
-      <EthereumProvider>
-        <div style={{ display: 'flex', flexDirection: 'column', gap: '10px' }}>
-          <DashboardHeader />
-          <DashboardBody />
-        </div>
-        <ToastContainer
-          position="top-right"
-          icon={false}
-          autoClose={5000}
-          hideProgressBar={false}
-          newestOnTop
-          closeOnClick
-          rtl={false}
-          pauseOnFocusLoss
-          draggable
-          pauseOnHover
-          progressStyle={{ backgroundColor: 'black' }}
-        />
-      </EthereumProvider>
-    </SafeHydrate>
+    <div style={{ display: 'flex', flexDirection: 'column', gap: '10px' }}>
+      <ConnectSection />
+      <div className="Dashboard">
+        <DashboardHeader />
+        <DashboardBody />
+      </div>
+    </div>
   );
 }
