+++ components/Dashboard/TokenBalance.tsx
@@ -1,21 +1,25 @@
+import { useEthereum } from 'lib/hooks/useEthereum';
+import { Erc20TokenData, TokenData } from 'lib/interfaces';
 import { getBalanceText } from 'lib/utils';
+import { getChainExplorerUrl } from 'lib/utils/chains';
 import TokenLogo from '../common/TokenLogo';
 
 interface Props {
-  symbol: string;
-  icon: string;
-  balance: string;
-  decimals?: number;
-  explorerUrl: string;
+  token: TokenData;
 }
 
-const TokenBalance = ({ symbol, icon, balance, decimals, explorerUrl }: Props) => (
-  <div className="TokenBalance">
-    <a href={explorerUrl} style={{ color: 'black' }}>
-      <TokenLogo src={icon} alt={symbol} />
-      {getBalanceText(symbol, balance, decimals)}
-    </a>
-  </div>
-);
+const TokenBalance = ({ token }: Props) => {
+  const { selectedChainId } = useEthereum();
+  const explorerUrl = `${getChainExplorerUrl(selectedChainId)}/address/${token.contract.address}`;
+
+  return (
+    <div className="TokenBalance">
+      <a href={explorerUrl} style={{ color: 'black' }}>
+        <TokenLogo src={token.icon} alt={token.symbol} />
+        {getBalanceText(token.symbol, token.balance, (token as Erc20TokenData).decimals)}
+      </a>
+    </div>
+  );
+};
 
 export default TokenBalance;
