+++ lib/utils/erc721.ts
@@ -1,16 +1,9 @@
 import { BigNumber, Contract, providers, utils } from 'ethers';
-import { getAddress } from 'ethers/lib/utils';
 import { OPENSEA_REGISTRY } from 'lib/abis';
-import {
-  ADDRESS_ZERO,
-  DUMMY_ADDRESS,
-  DUMMY_ADDRESS_2,
-  MOONBIRDS_ADDRESS,
-  OPENSEA_REGISTRY_ADDRESS,
-} from 'lib/constants';
-import { IERC721Allowance, TokenMapping } from 'lib/interfaces';
-import { shortenAddress, topicToAddress } from '.';
-import { convertString, unpackResult, withFallback } from './promises';
+import { ADDRESS_ZERO, MOONBIRDS_ADDRESS, OPENSEA_REGISTRY_ADDRESS } from 'lib/constants';
+import { IERC721Allowance } from 'lib/interfaces';
+import { topicToAddress } from '.';
+import { unpackResult } from './promises';
 
 export async function getLimitedAllowancesFromApprovals(contract: Contract, approvals: providers.Log[]) {
   const deduplicatedApprovals = approvals.filter(
@@ -78,19 +71,6 @@ async function getUnlimitedAllowanceFromApproval(
   return { spender };
 }
 
-export async function getTokenData(contract: Contract, ownerAddress: string, tokenMapping: TokenMapping = {}) {
-  const tokenData = tokenMapping[getAddress(contract.address)];
-
-  const [balance, symbol] = await Promise.all([
-    withFallback(convertString(unpackResult(contract.functions.balanceOf(ownerAddress))), 'ERC1155'),
-    // Use the tokenlist name if present, fall back to '???' since not every NFT has a name
-    tokenData?.name ?? withFallback(unpackResult(contract.functions.name()), shortenAddress(contract.address)),
-    throwIfNotErc721(contract),
-  ]);
-
-  return { symbol, balance };
-}
-
 export async function getOpenSeaProxyAddress(
   userAddress: string,
   provider: providers.Provider
@@ -137,15 +117,4 @@ export function generatePatchedAllowanceEvents(
       ],
     },
   ];
-}
-
-async function throwIfNotErc721(contract: Contract) {
-  // If the function isApprovedForAll does not exist it will throw (and is not ERC721)
-  const [isApprovedForAll] = await contract.functions.isApprovedForAll(DUMMY_ADDRESS, DUMMY_ADDRESS_2);
-
-  // The only acceptable value for checking whether 0x00...01 has an allowance set to 0x00...02 is false
-  // This could happen when the contract is not ERC721 but does have a fallback function
-  if (isApprovedForAll !== false) {
-    throw new Error('Response to isApprovedForAll was not false, indicating that this is not an ERC721 contract');
-  }
 }
